huge home display and divide each file
